+++ app/Http/Controllers/TarefaController.php
@@ -17,37 +17,73 @@ class TarefaController extends Controller
     public function store(Request $request)
     {
         $tarefa = new Tarefa();
-        $due_date = $request->due_date;
         $assignee_id = $request->assignee_id;
 
-        if ($tarefa->isValidDate($due_date) && $tarefa->isForeignId($assignee_id)) {
-            Tarefa::Create($request->all());
-        } else {
-            echo "Dados inválidos. \nO departamento deve existir. \nA data pode estar inválida ou não está no formato correto.";
-        }
+        try {
+            if (!$tarefa->isForeignId($assignee_id)) {
+                return "Funcionário inválido";
+            }
+            return Tarefa::create($request->all());
 
+        } catch (\Illuminate\Database\QueryException $e) {
+            if ($e->getCode() === '22001') {
+                return response()->json(['status' => 'Limite de caracteres atingido']);
+            }
+            if ($e->getCode() === 'HY000') {
+                return response()->json(['status' => 'Campo faltando']);
+            }
+            if($e->getCode() === '22007'){
+                return response()->json(['status' => 'Formato da data inconrreto. Tente: yyyy-mm-dd hh:mm:ss']);
+            }
+        }
     }
 
     public function show(string $id)
     {
-        return Tarefa::findOrFail($id);
+        $tarefa = new Tarefa();
+
+        if ($tarefa->existId($id)) {
+            return Tarefa::findOrFail($id);
+        } else {
+            return "Não encontrado.";
+        }
     }
     public function update(Request $request, string $id)
     {
         $tarefa = new Tarefa();
-        $tarefa::findOrFail($id);
-        $due_date = $request->due_date;
+        $assignee_id = $request->assignee_id;
 
-        if ($tarefa->isValidDate($due_date)) {
-            $tarefa->update($request->all());
-        } else {
-            echo "Dados inválidos. \nO departamento deve existir. \nA data pode estar inválida ou não está no formato correto.";
+        try {
+            if (!$tarefa->existId($id)) {
+                return "Tarefa não encontrada";
+            }
+            if ($tarefa->isForeignId($assignee_id)) {
+                return "Funcionário inválido";
+            }
+
+            $tarefa = Tarefa::findOrFail($id);
+            return $tarefa->update($request->all());
+
+        } catch (\Illuminate\Database\QueryException $e) {
+            if ($e->getCode() === '22001') {
+                return response()->json(['status' => 'Limite de caracteres atingido']);
+            }
+            if($e->getCode() === '22007'){
+                return response()->json(['status' => 'Formato da data inconrreto. Tente: yyyy-mm-dd hh:mm:ss']);
+            }
         }
 
     }
     public function destroy(string $id)
     {
-        $tarefa = Tarefa::findOrFail($id);
-        $tarefa->delete();
+        $tarefa = new Tarefa();
+
+        if ($tarefa->existId($id)) {
+            $tarefa = Tarefa::findOrFail($id);
+            $tarefa->delete();
+        } else {
+            return "Tarefa não encontrada.";
+        }
+
     }
 }
